requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
drf-nested-routers==0.94.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
